projects/hotnews/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic==2.10.6
httpx==0.28.1
feedparser==6.0.11
python-dateutil==2.9.0.post0
